requirements.txt
twilio==6.63.2
djangorestframework==3.9.2
coreapi==2.3.3
gunicorn==21.2.0
Django==2.2.28
requests==2.31.0
django_extensions
typing_extensions

# testing
pytest
pytest-django